requirements.txt

google-generativeai>=0.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.13
beautifulsoup4>=4.12.0
requests>=2.31.0
#psycopg2-binary>=2.9.9

python-dotenv==1.2.1
google-genai==1.55.0
streamlit==1.52.0



yt-dlp>=2023.10.13

pymupdf>=1.23.0
pypdf>=3.17.0

reportlab>=4.0.4
protobuf~=5.29.5
psycopg2-binary~=2.9.11

